Add user end dependencies to the main file
