+++ er-labrador/.env
@@ -1,0 +1,3 @@
+HOST=uom.eratos
+ESLINT=1
+PORT=3000
