+++ er-labrador/package.json
@@ -1,6 +1,6 @@
 {
-	"name": "er-labrador3",
-	"version": "0.1.0",
+	"name": "eratos-webapp",
+	"version": "3.1.1",
 	"private": true,
 	"dependencies": {
 		"@ant-design/icons": "^4.6.2",
@@ -21,8 +21,28 @@
 		"react-router-dom": "^5.2.0",
 		"react-scripts": "4.0.3",
 		"redux": "^4.0.5",
-		"web-vitals": "^1.0.1"
+		"web-vitals": "^1.0.1",
+		"@auth0/auth0-spa-js": "^1.15.0",
+		"@stripe/react-stripe-js": "^1.4.0",
+		"@stripe/stripe-js": "^1.15.0",
+		"cors-anywhere": "^0.4.4",
+		"element-react": "^1.4.34",
+		"element-theme-default": "^1.4.13",
+		"fetch-jsonp": "^1.1.3",
+		"mapbox-gl": "^2.2.0",
+		"mapbox-gl-draw": "^0.16.0",
+		"mapbox-gl-geocoder": "^2.0.1",
+		"moment": "^2.29.1",
+		"pubsub-js": "^1.9.3",
+		"pubsubjs": "^0.3.9",
+		"querystring": "^0.2.1",
+		"react-payment-inputs": "^1.1.8",
+		"start": "^5.1.0",
+		"styled-components": "^5.2.3",
+		"turf": "^3.0.14",
+		"url-safe-base64": "^1.1.1"
 	},
+	"proxy": "http://localhost:4242",
 	"scripts": {
 		"start": "cross-env HTTPS=true HOST=uom.eratos webpack-dev-server --mode development --inline --hot --open",
 		"build": "webpack --mode production",
@@ -70,6 +90,7 @@
 		"webpack": "^5.24.3",
 		"webpack-bundle-analyzer": "^4.4.0",
 		"webpack-cli": "^3.3.12",
-		"webpack-dev-server": "^3.11.2"
+		"webpack-dev-server": "^3.11.2",
+		"fetch-mock": "^9.11.0"
 	}
 }
