+++ er-labrador/src/components/Admin/moduleDetail.js
@@ -1,12 +1,12 @@
 import React, { Component } from "react";
 import { connect } from "react-redux";
 import { withAuth0 } from "@auth0/auth0-react";
-import { UPDATE_USER } from "../store/actionTypes";
+import { UPDATE_USER } from "../../store/actionTypes";
 
 import NavBar from "./navBar";
-import "../styles/index.css";
+import "../../styles/index.css";
 import { Descriptions, Button, message } from "antd";
-import { createModifyModule } from "../store/api";
+import { createModifyModule } from "../../store/api";
 import axios from "axios";
 
 import ProForm, {
