+++ er-labrador/src/components/Admin/userDetail.js
@@ -1,10 +1,10 @@
 import React, { Component } from "react";
 import { connect } from "react-redux";
 import { withAuth0 } from "@auth0/auth0-react";
-import { UPDATE_USER } from "../store/actionTypes";
+import { UPDATE_USER } from "../../store/actionTypes";
 
 import NavBar from "./navBar";
-import "../styles/index.css";
+import "../../styles/index.css";
 import { Descriptions, Table, Input, Button, Space } from "antd";
 import { Button as RBButton } from "react-bootstrap";
 
@@ -15,15 +15,13 @@ class UserDetail extends Component {
 			searchText: "",
 			searchedColumn: "",
 			currRow: props.location.state.currRow,
-			size: 'default',
-		};	
-
+			size: "default",
+		};
 	}
 
 	componentDidMount = () => {
 		const { user } = this.props.auth0;
 		if (this.props.user !== user) this.props.updateUser(user);
-		
 	};
 
 	render() {
@@ -32,20 +30,35 @@ class UserDetail extends Component {
 				<NavBar defaultSelectedKeys="3">
 					<div className="user-detail-content common-component-content">
 						<Descriptions
-          					bordered
-          					title="User Details"
-          					size={this.state.size}
-        				>
-          				<Descriptions.Item label="User ID">{this.state.currRow.UserID}</Descriptions.Item>
-          				<Descriptions.Item label="Full Name">{this.state.currRow.Name}</Descriptions.Item>
-          				<Descriptions.Item label="Email">{this.state.currRow.Email}</Descriptions.Item>
-          				<Descriptions.Item label="Eratos User ID">{this.state.currRow.EratosUserID}</Descriptions.Item>
-          				<Descriptions.Item label="Auth0 ID">{this.state.currRow.Auth0ID}</Descriptions.Item>
-          				<Descriptions.Item label="Administrator">{this.state.currRow.isAdmin}</Descriptions.Item>
-          				<Descriptions.Item label="Date and Time of User Created">{this.state.currRow.CreatedAt}</Descriptions.Item>
-          				<Descriptions.Item label="Info">{this.state.currRow.Info}</Descriptions.Item>
-            			</Descriptions>
-
+							bordered
+							title="User Details"
+							size={this.state.size}
+						>
+							<Descriptions.Item label="User ID">
+								{this.state.currRow.UserID}
+							</Descriptions.Item>
+							<Descriptions.Item label="Full Name">
+								{this.state.currRow.Name}
+							</Descriptions.Item>
+							<Descriptions.Item label="Email">
+								{this.state.currRow.Email}
+							</Descriptions.Item>
+							<Descriptions.Item label="Eratos User ID">
+								{this.state.currRow.EratosUserID}
+							</Descriptions.Item>
+							<Descriptions.Item label="Auth0 ID">
+								{this.state.currRow.Auth0ID}
+							</Descriptions.Item>
+							<Descriptions.Item label="Administrator">
+								{this.state.currRow.isAdmin}
+							</Descriptions.Item>
+							<Descriptions.Item label="Date and Time of User Created">
+								{this.state.currRow.CreatedAt}
+							</Descriptions.Item>
+							<Descriptions.Item label="Info">
+								{this.state.currRow.Info}
+							</Descriptions.Item>
+						</Descriptions>
 					</div>
 				</NavBar>
 			</div>
