+++ er-labrador/src/components/Admin/userProfile.js
@@ -2,17 +2,17 @@ import React, { Component } from "react";
 import { Route, Redirect } from "react-router";
 import { connect } from "react-redux";
 import { withAuth0 } from "@auth0/auth0-react";
-import { UPDATE_USER } from "../store/actionTypes";
+import { UPDATE_USER } from "../../store/actionTypes";
 
 import NavBar from "./navBar";
-import "../styles/index.css";
+import "../../styles/index.css";
 import { Input, Button, Switch, Form, message } from "antd";
 import {
 	REACT_APP_ERATOS_TRACKER,
 	REACT_APP_ERATOS_AUTH0_AUD,
-} from "../store/auth0";
+} from "../../store/auth0";
 import axios from "axios";
-import { updateUserInfo } from "../store/api";
+import { updateUserInfo } from "../../store/api";
 
 class UserProfile extends Component {
 	constructor(props) {
