+++ er-labrador/webpack.config.js
@@ -11,7 +11,7 @@ module.exports = {
 	mode: "development",
 	// devtool: "cheap-module-source-map",
 	devServer: {
-		hot: true, //在devServer中增加hot字段
+		hot: true, //import Hot to the dev server
 		inline: true,
 		contentBase: path.join(__dirname, "./dist"),
 		publicPath: "/",
@@ -25,21 +25,12 @@ module.exports = {
 		},
 		historyApiFallback: true,
 	},
-	// entry: ["./src/index.js", "./src/dev.js"], //在entry字段中添加触发文件配置
-
-	// entry: {
-	// 	hot: "webpack/hot/dev-server",
-	// 	index: path.resolve(__dirname, "./src/index.js"),
-	// },
 	entry: {
 		index: path.join(__dirname, "./src/index.js"),
 	},
 	// devtool: "inline-source-map",
 	output: {
-		// 输出路径
-		// __dirname nodejs的变量，代表当前文件的目录绝对路径
 		path: path.join(__dirname + "/dist"),
-		// 输出文件名
 		filename: "[name].bundle.min.js",
 		chunkFilename: "[name].[chunkhash:4].child.js",
 	},
@@ -47,16 +38,13 @@ module.exports = {
 	module: {
 		rules: [
 			{
-				test: /\.(js|jsx)?$/, // jsx文件的正则
-				exclude: /node_modules/, // 排除 node_modules 文件夹
+				test: /\.(js|jsx)?$/,
+				exclude: /node_modules/,
 				use: {
-					// loader 是 babel
 					loader: "babel-loader",
 					options: {
-						// babel 转义的配置选项
 						babelrc: false,
 						presets: [
-							// 添加 preset-react
 							require.resolve("@babel/preset-react"),
 							[
 								require.resolve("@babel/preset-env"),
@@ -69,30 +57,14 @@ module.exports = {
 				},
 			},
 			{
-				// 写正则表达式，匹配哪些文件
 				test: /\.css$/,
-				// 使用哪些loader进行处理
-				use: [
-					// use数组中loader执行顺序：从右到左，从下到上 依次执行
-					// 创建style标签，将js中的样式资源插入进去，添加到head中生效
-					"style-loader",
-					// 将css文件变成commonJS模块加载到js中，里面内容是样式字符串
-					"css-loader",
-					"resolve-url-loader",
-				],
+				use: ["style-loader", "css-loader", "resolve-url-loader"],
 			},
 			{
 				test: /\.less$/,
-				use: [
-					"style-loader",
-					"css-loader",
-					// 将less文件变成css文件
-					// 需要下载 less-loader和less
-					"less-loader",
-				],
+				use: ["style-loader", "css-loader", "less-loader"],
 			},
 			{
-				// test: /\.(jpg|png|gif|svg|jpeg)?$/,
 				test: /\.(gif|jpg|png|woff|svg|eot|ttf)$/,
 				use: [
 					"url-loader?limit=8192&name=ui/[name].[hash].[ext]!extract-loader!file-loader",
@@ -102,7 +74,6 @@ module.exports = {
 	},
 	plugins: [
 		// new BundleAnalyzerPlugin(),
-		// plugins中增加下面内容，实例化热加载插件
 		// new webpack.HotModuleReplacementPlugin(),
 		new HtmlWebPackPlugin({
 			template: "public/index.html",
@@ -131,7 +102,6 @@ module.exports = {
 			},
 		}),
 	],
-	// 将 jsx 添加到默认扩展名中，省略 jsx
 	resolve: {
 		extensions: [".wasm", ".mjs", ".js", ".json", ".jsx"],
 	},
